Improve overall compilation stats after the telescope picker feature
