+++ crates/file_finder/Cargo.toml
@@ -27,6 +27,7 @@ language.workspace = true
 menu.workspace = true
 open_path_prompt.workspace = true
 picker.workspace = true
+picker_preview.workspace = true
 project.workspace = true
 settings.workspace = true
 serde.workspace = true
