+++ crates/picker_preview/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "picker_preview"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/picker_preview.rs"
+doctest = false
+
+[dependencies]
+anyhow.workspace = true
+editor.workspace = true
+gpui.workspace = true
+language.workspace = true
+multi_buffer.workspace = true
+picker.workspace = true
+project.workspace = true
+rope.workspace = true
+settings.workspace = true
+ui.workspace = true
+util.workspace = true
