+++ crates/search/src/text_finder.rs
@@ -265,7 +265,8 @@ impl TextFinder {
         cx: &mut Context<Self>,
     ) -> Self {
         let project = delegate.project(cx).clone();
-        let picker = cx.new(|cx| Picker::list_with_preview(delegate, project, window, cx));
+        let preview = picker_preview::editor_preview(project, window, cx);
+        let picker = cx.new(|cx| Picker::list_with_preview(delegate, preview, window, cx));
         let picker_weak = picker.downgrade();
         let picker_focus_handle = picker.focus_handle(cx);
         picker.update(cx, |picker, cx| {
